Major changes.
Now based on the newer (more complicated) micro-bunzip that is part of the
current version of busybox. The original code is in `micro-bunzip.orig.c`
for reference.

I modified it so that it allows external control over decompressing blocks,
rather than `read_bunzip` automatically filling the buffer, you have to do
that for it (it will return 0 bytes decoded, you then init the next block,
checking for the EOF block).

This allows for what we want to do (determining exactly where blocks are and
seeking to/decompressing a single block without waste -- the previous version
would have to suck another block from the file for each block desired).

bzip-table now uncompresses the entire block and reports the length in bytes.
Why? Because it has to, bzip2 blocks are NOT ALL THE SAME SIZE, despite
many reports on the web. Blocks are exactly bs * 100,000k AFTER run length
encoding, but before RLE they can be fairly arbitrary chunks of the input.

Everything seems to work (see make test). This is not ideal because of the
block sizes, we'll have to use a search rather than a lookup table to find
the block corresponding to a byte in the uncompressed file. But, it will
work -- and the limitations help the argument for a new file format.